+++ src/app/about/page.tsx
@@ -5,7 +5,7 @@ import { TextBtn } from '@/features/ui/text-btn'
 import { Title } from '@/features/ui/title'
 import { AboutSnsBox } from './components/about-sns-box'
 
-import { setMetaData } from '@/utils'
+import { setMetaData } from '@/utils/set-metadata'
 
 const description = 'このサイトについて'
