+++ src/app/articles/components/articles-body.tsx
@@ -2,7 +2,7 @@ import { Title } from '@/features/ui/title'
 import { TextBtn } from '@/features/ui/text-btn'
 import { ArticlesList } from './articles-list'
 
-import { FeedObj } from '@/types'
+import { FeedObj } from '@/types/feed'
 
 type Props = {
   articles: FeedObj[]
