+++ src/app/articles/components/articles-list.tsx
@@ -1,7 +1,7 @@
 import { CategoryIcon } from '@/features/ui/category-icon'
 import { ImageWrapper } from '@/features/ui/image-wrapper'
 
-import { FeedObj } from '@/types'
+import { FeedObj } from '@/types/feed'
 
 type Props = {
   articles: FeedObj[]
