--- src/app/contact/libs/index.ts
@@ -1,2 +1,0 @@
-export { sendMail } from './send-mail'
-export { verifyRecaptcha } from './verify-recaptcha'
