+++ src/app/error.tsx
@@ -1,7 +1,7 @@
 'use client'
 import { useEffect } from 'react'
 
-import { setMetaData } from '@/utils'
+import { setMetaData } from '@/utils/set-metadata'
 
 const description = 'ページでエラーが発生しました'
