--- src/hooks/index.ts
@@ -1,2 +1,0 @@
-export * from './use-debug-mode'
-export * from './use-page-view'
