+++ src/libs/fetch/fetch-works-index.ts
@@ -3,10 +3,10 @@
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 import { MicroCMSListResponse } from 'microcms-js-sdk'
 
-import { microcmsClient } from '@/libs'
-import { WORKS_PER_PAGE, CATEGORY_RIMIT } from '@/const'
+import { microcmsClient } from './microcms-client'
+import { WORKS_PER_PAGE, CATEGORY_RIMIT } from '@/const/microcms'
 
-import type { WorkIndex, WorksCategory } from '@/types'
+import type { WorkIndex, WorksCategory } from '@/types/microcms'
 
 export async function fetchWorksIndex() {
   const result = await Promise.allSettled([
