--- src/libs/fetch/index.ts
@@ -1,5 +1,0 @@
-export { fetchArticles } from './fetch-articles'
-export { fetchTopList } from './fetch-top-list'
-export { fetchWorksDetail } from './fetch-works-detail'
-export { fetchWorksIndex } from './fetch-works-index'
-export { microcmsClient } from './microcms-client'
