--- src/types/index.ts
@@ -1,2 +1,0 @@
-export * from './feed'
-export * from './microcms'
