--- src/utils/index.ts
@@ -1,7 +1,0 @@
-export { conversionDate } from './conversion-date'
-export { cutText } from './cut-text'
-export { isEmptyObject } from './is-empty-object'
-export { loadScript } from './load-script'
-export { removeHtml } from './remove-html'
-export { setMetaData } from './set-metadata'
-export { shuffle } from './shuffle'
